+++ part4/blog-list/src/test/blog_api.test.js
@@ -5,7 +5,18 @@ const api = supertest(app);
 
 const Blog = require("../models/blog");
 const { initialBlogs, oneBlog } = require("./mock_blogs");
+const { createUser, blogsInDb, delAllUser } = require("./test_helpers");
 const url = "/api/blogs";
+let token = "Bearer ";
+beforeAll(async () => {
+  await delAllUser();
+  await createUser("root", "root", "sekret");
+  const result = await api
+    .post("/api/login")
+    .send({ username: "root", password: "sekret" });
+
+  token += result.body.token;
+});
 
 beforeEach(async () => {
   await Blog.deleteMany({});
@@ -18,19 +29,23 @@ beforeEach(async () => {
 
 describe("when there is some blogs", () => {
   test("all blogs are returned", async () => {
-    const res = await api.get(url).expect(200);
+    const res = await api.get(url).set("Authorization", token).expect(200);
     expect(res.body).toHaveLength(initialBlogs.length);
   });
 
   test("blogs are returned as json", async () => {
     await api
       .get("/api/blogs")
+      .set("Authorization", token)
       .expect(200)
       .expect("Content-Type", /application\/json/);
   });
 
   test("blogs are defined", async () => {
-    const res = await api.get("/api/blogs").expect(200);
+    const res = await api
+      .get("/api/blogs")
+      .set("Authorization", token)
+      .expect(200);
 
     const content = res.body.map((e) => e.id);
     content.forEach((element) => {
@@ -41,14 +56,11 @@ describe("when there is some blogs", () => {
 
 describe("addition of a new blog", () => {
   test("post a blog", async () => {
-    await api
-      .post(url)
-      .send(oneBlog)
-      .expect(201)
-      .expect("Content-Type", /application\/json/);
+    const blogsAtStart = await blogsInDb();
+    await api.post(url).set("Authorization", token).expect(201).send(oneBlog);
+    const blogsAtEnd = await blogsInDb();
 
-    const blogs = await Blog.find({});
-    expect(blogs).toHaveLength(initialBlogs.length + 1);
+    expect(blogsAtEnd).toHaveLength(blogsAtStart.length + 1);
   });
 
   test("post blogs without likes", async () => {
@@ -56,6 +68,7 @@ describe("addition of a new blog", () => {
     delete newBlog.likes;
     const res = await api
       .post(url)
+      .set("Authorization", token)
       .send(newBlog)
       .expect(201)
       .expect("Content-Type", /application\/json/);
@@ -68,6 +81,7 @@ describe("addition of a new blog", () => {
     delete newBlog.url;
     const res = await api
       .post(url)
+      .set("Authorization", token)
       .send(newBlog)
       .expect(400)
       .expect("Content-Type", /application\/json/);
@@ -77,12 +91,13 @@ describe("addition of a new blog", () => {
 
 describe("update of a blog", () => {
   test("update blog", async () => {
-    const res = await api.post(url).send(oneBlog);
+    const res = await api.post(url).set("Authorization", token).send(oneBlog);
     const id = res.body.id;
     const title = "im updated";
     const updatedBlog = { title };
     const updResult = await api
       .put(`${url}/${id}`)
+      .set("Authorization", token)
       .send(updatedBlog)
       .expect(200);
     expect(updResult.body.title).toContain(title);
@@ -91,9 +106,9 @@ describe("update of a blog", () => {
 
 describe("delete of a blog", () => {
   test("delete blog", async () => {
-    const res = await api.post(url).send(oneBlog);
+    const res = await api.post(url).set("Authorization", token).send(oneBlog);
     const id = res.body.id;
-    await api.delete(`${url}/${id}`).expect(204);
+    await api.delete(`${url}/${id}`).set("Authorization", token).expect(204);
   });
 });
